move cli into migrator package
